Day9 (#9)
* Convert machine to longs

* Part 1 solution

* Part 2 solution

+++ 2019/2/Utils.cs
@@ -1,10 +1,10 @@
 namespace aoc.y2019.day2
 {
-    abstract class Solver : aoc.utils.ProblemSolver<int>
+    abstract class Solver : aoc.utils.ProblemSolver<long>
     {
         protected Solver(string file, int exp) : base(file, exp) { }
 
-        protected int runMachine(int[] prog, int noun, int verb) {
+        protected long runMachine(long[] prog, int noun, int verb) {
             var mach = new aoc.intcode.Machine(prog);
 
             mach[1] = noun;
