+++ 2019/7/Part2.cs
@@ -6,8 +6,8 @@ namespace aoc.y2019.day7
     {
         public Part2(string file, int exp) : base(file, exp) { }
 
-        protected override int runChain(List<Amp> chain) {
-            int? signal = 0;
+        protected override long runChain(List<Amp> chain) {
+            long? signal = 0;
             var lastSignal = 0;
 
             for (var ndx = 0; !chain[chain.Count - 1].mach.isHalted(); ndx = (ndx + 1) % chain.Count) {
@@ -34,7 +34,7 @@ namespace aoc.y2019.day7
             return lastSignal;
         }
 
-        protected override int doWork() {
+        protected override long doWork() {
             return base.doWork(new List<int>() { 5, 6, 7, 8, 9 });
         }
     }
