+++ intcode/Parser.cs
@@ -4,12 +4,12 @@ namespace aoc.intcode
 {
     class Parser
     {
-        public static int[] parseInput(string fileName) {
+        public static long[] parseInput(string fileName) {
             var lines = aoc.utils.Parser.readLines(fileName);
             var line = lines[0];
             var ints = line.Split(',');
 
-            return Array.ConvertAll(ints, n => int.Parse(n));
+            return Array.ConvertAll(ints, n => long.Parse(n));
         }
     }
 }
